moving an upload modal into the settings page
