+++ client/src/App.js
@@ -26,7 +26,7 @@ import NewRestaurant from './components/Container/CreateRestaurantPage';
 import RequireNotAuth from './components/HOC/RequireNotAuth';
 import RequireAuth from './components/HOC/RequireAuth';
 import { sidebar } from './config/conditionalPathnames';
-import Test from './components/Presentational/Test';
+import UploadModal from './components/Presentational/UploadModal';
 
 const AuthedPartyPage = RequireAuth(PartyPage);
 const AuthedLoginEmployee = RequireAuth(LoginEmployee, true);
@@ -75,7 +75,7 @@ class App extends Component {
               push={history.push}
             />
             <Switch>
-              <Route path="/test" component={Test} />
+              <Route path="/test" component={UploadModal} />
               <Route path="/" component={Landing} exact />
               <Route path="/logout" component={Logout} />
               <Route path="/login" component={NotAuthedLogin} />
